BackEnd Java

+++ Backend/src/main/java/com/example/EmployeeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RestResource;
 
 
-public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Long> {
+public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Integer> {
 
 		public List<Employee> findByLastNameOrFirstName(@Param("last_name")String lastName,@Param("first_name")String firstName);
